extract form into CreateFlightForm component
